Move mode definitions to modes.ts
