+++ src/server.ts
@@ -3,6 +3,7 @@ import chalk from "chalk";
 import http from "http";
 import { ensureBuffer } from "./buffer";
 import { findNextRecordToReplay, findRecordMatches } from "./matcher";
+import { Mode } from "./modes";
 import { Persistence } from "./persistence";
 import { proxy } from "./proxy";
 import { TapeRecord } from "./tape";
@@ -363,28 +364,3 @@ function extractPath(url: string) {
 }
 
 const DEFAULT_TAPE = "default";
-
-/**
- * Possible modes.
- */
-export type Mode = ReplayMode | RecordMode | MimicMode | PassthroughMode;
-
-/**
- * Replays requests from tapes. Fails any unexpected requests.
- */
-export type ReplayMode = "replay";
-
-/**
- * Records requests. Ignores recorded tapes.
- */
-export type RecordMode = "record";
-
-/**
- * Records requests the first time it encounters them, then replays them.
- */
-export type MimicMode = "mimic";
-
-/**
- * Acts as a pass-through proxy. No recording occurs.
- */
-export type PassthroughMode = "passthrough";
